fixed deposit

+++ phase1/src/main/java/atm/CreditCard.java
@@ -3,7 +3,7 @@ package atm;
 import java.text.DecimalFormat;
 
 public class CreditCard extends Account {
-    private int balance;
+    private double balance;
     private DecimalFormat currencyFormat = new DecimalFormat("0.00");
 
     public CreditCard(User owner) {
